requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.1.0
cachelib==0.9.0
click==8.1.3
Deprecated==1.2.13
Faker==13.15.1
flasgger==0.9.5
Flask==2.1.3
Flask-Caching==2.0.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
mistune==2.0.4
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
sa-filters==1.1.0
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
Werkzeug==2.2.1
wrapt==1.14.1
zipp==3.8.1
